Sword catching game
The objective of the game is to stop a sword before it hits you. If you do so, you are awarded a point, if the sowrd hits you then the other player is awarded a point. Players switch roles half way through the game. There are some other mechanics that are explained in the comments of the code or in the instructions.

The program works as I intended it to however it is extremley laggy and I am not sure why.